feat: Update note functionality to include title
The update note functionality has been modified to include updating the note's title in addition to its content. The `updateNote` method in the `DriveRemoteDataSource` now accepts a title parameter, which is used to update the file's name in Google Drive. Corresponding changes have been made in the repository, use case, and presentation layers to pass the title along with the content when updating a note.